fix: validation location
